+++ internal/css_ast/css_ast.go
@@ -43,6 +43,9 @@ type Token struct {
 	// This index points to that import record.
 	ImportRecordIndex uint32 // 4 bytes
 
+	// The division between the number and the unit for "TDimension" tokens.
+	UnitOffset uint16 // 2 bytes
+
 	// This will never be "TWhitespace" because whitespace isn't stored as a
 	// token directly. Instead it is stored in "HasWhitespaceAfter" on the
 	// previous token.
@@ -52,6 +55,14 @@ type Token struct {
 	// token. This isn't strictly true in some cases because sometimes this flag
 	// is changed to make the generated code look better (e.g. around commas).
 	HasWhitespaceAfter bool // 1 byte
+}
+
+func (t Token) DimensionValue() string {
+	return t.Text[:t.UnitOffset]
+}
+
+func (t Token) DimensionUnit() string {
+	return t.Text[t.UnitOffset:]
 }
 
 // This interface is never called. Its purpose is to encode a variant type in